refactor: to-be-time -> time으로 변경
